Dedupe formatSeconds, fixing the "1m 60s" bug on the health view
PR #1144 fixed the minute-rounding bug in the studio copy of
formatSeconds but left GameHealthView's own copy behind, so the
operator health table still rendered "1m 60s" for e.g. 119.6s.

Move the fixed implementation to relativeTime.ts — which already
owns formatRelativeTime and formatDuration — and point both the
studio panels and GameHealthView at it. The formatSeconds tests
move alongside it.
