+++ eslint-rules/module-size-baseline.json
@@ -671,7 +671,7 @@
     "apps/web/src/recentPlays.ts": 36,
     "apps/web/src/recommendationsApi.test.ts": 71,
     "apps/web/src/recommendationsApi.ts": 197,
-    "apps/web/src/relativeTime.ts": 41,
+    "apps/web/src/relativeTime.ts": 50,
     "apps/web/src/remixApi.ts": 226,
     "apps/web/src/remixChatCopy.test.ts": 27,
     "apps/web/src/remixChatCopy.ts": 52,
